+++ src/test/kotlin/com/zepben/evolve/services/network/tracing/phases/PhaseSelectorTest.kt
@@ -7,8 +7,12 @@
  */
 package com.zepben.evolve.services.network.tracing.phases
 
+import com.zepben.evolve.cim.iec61970.base.core.PhaseCode
 import com.zepben.evolve.cim.iec61970.base.core.Terminal
 import com.zepben.evolve.cim.iec61970.base.wires.SinglePhaseKind
+import org.hamcrest.MatcherAssert.assertThat
+import org.hamcrest.Matchers.equalTo
+import org.hamcrest.Matchers.nullValue
 import org.junit.jupiter.api.Test
 import org.mockito.kotlin.doReturn
 import org.mockito.kotlin.mock
@@ -75,6 +79,83 @@ class PhaseSelectorTest {
         verify(current)[SinglePhaseKind.Y] = SinglePhaseKind.N
 
         verifyDone()
+    }
+
+    @Test
+    fun testAsPhaseCodesThree() {
+        val terminal = Terminal().apply { phases = PhaseCode.ABCN }
+        val normalPhases = terminal.normalPhases
+        val currentPhases = terminal.currentPhases
+
+        assertThat(normalPhases.asPhaseCode(), equalTo(PhaseCode.NONE))
+        assertThat(currentPhases.asPhaseCode(), equalTo(PhaseCode.NONE))
+
+        normalPhases[SinglePhaseKind.A] = SinglePhaseKind.A
+
+        assertThat(normalPhases.asPhaseCode(), nullValue())
+        assertThat(currentPhases.asPhaseCode(), equalTo(PhaseCode.NONE))
+
+        currentPhases[SinglePhaseKind.A] = SinglePhaseKind.A
+
+        assertThat(normalPhases.asPhaseCode(), nullValue())
+        assertThat(currentPhases.asPhaseCode(), nullValue())
+
+        normalPhases[SinglePhaseKind.B] = SinglePhaseKind.B
+        normalPhases[SinglePhaseKind.C] = SinglePhaseKind.C
+        normalPhases[SinglePhaseKind.N] = SinglePhaseKind.N
+
+        assertThat(normalPhases.asPhaseCode(), equalTo(PhaseCode.ABCN))
+        assertThat(currentPhases.asPhaseCode(), nullValue())
+
+        currentPhases[SinglePhaseKind.B] = SinglePhaseKind.B
+        currentPhases[SinglePhaseKind.C] = SinglePhaseKind.C
+        currentPhases[SinglePhaseKind.N] = SinglePhaseKind.N
+
+        assertThat(normalPhases.asPhaseCode(), equalTo(PhaseCode.ABCN))
+        assertThat(currentPhases.asPhaseCode(), equalTo(PhaseCode.ABCN))
+    }
+
+    @Test
+    fun testAsPhaseCodesSingle() {
+        val terminal = Terminal().apply { phases = PhaseCode.BC }
+        val normalPhases = terminal.normalPhases
+        val currentPhases = terminal.currentPhases
+
+        assertThat(normalPhases.asPhaseCode(), equalTo(PhaseCode.NONE))
+        assertThat(currentPhases.asPhaseCode(), equalTo(PhaseCode.NONE))
+
+        normalPhases[SinglePhaseKind.A] = SinglePhaseKind.A
+        currentPhases[SinglePhaseKind.A] = SinglePhaseKind.A
+
+        assertThat(normalPhases.asPhaseCode(), equalTo(PhaseCode.NONE))
+        assertThat(currentPhases.asPhaseCode(), equalTo(PhaseCode.NONE))
+
+        normalPhases[SinglePhaseKind.B] = SinglePhaseKind.B
+        currentPhases[SinglePhaseKind.B] = SinglePhaseKind.B
+
+        assertThat(normalPhases.asPhaseCode(), nullValue())
+        assertThat(currentPhases.asPhaseCode(), nullValue())
+
+        normalPhases[SinglePhaseKind.C] = SinglePhaseKind.C
+        currentPhases[SinglePhaseKind.C] = SinglePhaseKind.C
+
+        assertThat(normalPhases.asPhaseCode(), equalTo(PhaseCode.BC))
+        assertThat(currentPhases.asPhaseCode(), equalTo(PhaseCode.BC))
+    }
+
+    @Test
+    fun testAsPhaseCodesChangingTerminalPhases() {
+        val terminal = Terminal().apply { phases = PhaseCode.BC }
+        val normalPhases = terminal.normalPhases
+
+        normalPhases[SinglePhaseKind.A] = SinglePhaseKind.A
+        normalPhases[SinglePhaseKind.B] = SinglePhaseKind.B
+        normalPhases[SinglePhaseKind.C] = SinglePhaseKind.C
+
+        assertThat(normalPhases.asPhaseCode(), equalTo(PhaseCode.BC))
+
+        terminal.phases = PhaseCode.AC
+        assertThat(normalPhases.asPhaseCode(), equalTo(PhaseCode.AC))
     }
 
     private fun verifyDone() {