+++ changelog.md
@@ -10,6 +10,7 @@
 
 * Phases are now set through transformers allowing for phase changes. Only valid phase mappings are currently supported. If a valid mapping has been missed,
   please contact Zepben.
+* You can now translate `TracedPhases` to a `PhaseCode` if the individual phases make sense.
 
 ##### Fixes